Move the generic text-handling classes into `texty`.
