+++ src/texty/export/ComboText.js
@@ -1,8 +1,8 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { IntfText } from '#p/IntfText';
-import { TypeText } from '#p/TypeText';
+import { IntfText } from '#x/IntfText';
+import { TypeText } from '#x/TypeText';
 
 
 /**
