+++ src/texty/index.js
@@ -2,3 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from '#x/Chalk';
+export * from '#x/ComboText';
+export * from '#x/IndentedText';
+export * from '#x/IntfText';
+export * from '#x/StyledText';
+export * from '#x/TypeText';
